+++ controller/settingscontroller.php
@@ -2,7 +2,7 @@
 /**
  * @author Ascensio System SIA <integration@onlyoffice.com>
  *
- * (c) Copyright Ascensio System SIA 2024
+ * (c) Copyright Ascensio System SIA 2025
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -118,6 +118,7 @@ class SettingsController extends Controller {
 			"sameTab" => $this->config->getSameTab(),
 			"preview" => $this->config->getPreview(),
 			"cronChecker" => $this->config->getCronChecker(),
+			"emailNotifications" => $this->config->getEmailNotifications(),
 			"versionHistory" => $this->config->getVersionHistory(),
 			"protection" => $this->config->getProtection(),
 			"encryption" => $this->config->checkEncryptionModule(),
@@ -134,7 +135,8 @@ class SettingsController extends Controller {
 			"reviewDisplay" => $this->config->getCustomizationReviewDisplay(),
 			"theme" => $this->config->getCustomizationTheme(),
 			"templates" => $this->getGlobalTemplates(),
-			"linkToDocs" => $this->config->getLinkToDocs()
+			"linkToDocs" => $this->config->getLinkToDocs(),
+			"unknownAuthor" => $this->config->getUnknownAuthor()
 		];
 		return new TemplateResponse($this->appName, "settings", $data, "blank");
 	}
@@ -208,6 +210,7 @@ class SettingsController extends Controller {
 	 * @param bool $sameTab - open in the same tab
 	 * @param bool $preview - generate preview files
 	 * @param bool $cronChecker - disable cron checker
+	 * @param bool $emailNotifications - notifications via e-mail
 	 * @param bool $versionHistory - keep version history
 	 * @param array $limitGroups - list of groups
 	 * @param bool $chat - display chat
@@ -218,6 +221,7 @@ class SettingsController extends Controller {
 	 * @param bool $toolbarNoTabs - display toolbar tab
 	 * @param string $reviewDisplay - review viewing mode
 	 * @param string $theme - default theme mode
+	 * @param string $unknownAuthor - display unknown author
 	 *
 	 * @return array
 	 */
@@ -227,6 +231,7 @@ class SettingsController extends Controller {
 		$sameTab,
 		$preview,
 		$cronChecker,
+		$emailNotifications,
 		$versionHistory,
 		$limitGroups,
 		$chat,
@@ -236,13 +241,15 @@ class SettingsController extends Controller {
 		$help,
 		$toolbarNoTabs,
 		$reviewDisplay,
-		$theme
+		$theme,
+		$unknownAuthor
 	) {
 		$this->config->setDefaultFormats($defFormats);
 		$this->config->setEditableFormats($editFormats);
 		$this->config->setSameTab($sameTab);
 		$this->config->setPreview($preview);
 		$this->config->setCronChecker($cronChecker);
+		$this->config->setEmailNotifications($emailNotifications);
 		$this->config->setVersionHistory($versionHistory);
 		$this->config->setLimitGroups($limitGroups);
 		$this->config->setCustomizationChat($chat);
@@ -253,6 +260,7 @@ class SettingsController extends Controller {
 		$this->config->setCustomizationToolbarNoTabs($toolbarNoTabs);
 		$this->config->setCustomizationReviewDisplay($reviewDisplay);
 		$this->config->setCustomizationTheme($theme);
+		$this->config->setUnknownAuthor($unknownAuthor);
 
 		return [
 			];