+++ templates/settings.php
@@ -2,7 +2,7 @@
 /**
  * @author Ascensio System SIA <integration@onlyoffice.com>
  * 
- * (c) Copyright Ascensio System SIA 2024
+ * (c) Copyright Ascensio System SIA 2025
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ script("onlyoffice", "template");
 <div class="section section-onlyoffice">
 	<h2 id="onlyoffice">
 		ONLYOFFICE
-		<a target="_blank" class="icon-info svg" title="" href="https://api.onlyoffice.com/editors/owncloud" data-original-title="<?php p($l->t("Documentation")) ?>"></a>
+		<a target="_blank" class="icon-info svg" title="" href="https://helpcenter.onlyoffice.com/integration/gettingstarted-owncloud.aspx" data-original-title="<?php p($l->t("Documentation")) ?>"></a>
 	</h2>
 
 	<h2><?php p($l->t("Server settings")) ?></h2>
@@ -34,7 +34,6 @@ script("onlyoffice", "template");
 <?php if ($_["encryption"] === true) { ?>
 	<p class="onlyoffice-error">
 	<?php p($l->t("Encryption App is enabled, the application cannot work. You can continue working with the application if you enable master key.")) ?>
-	<a target="_blank" class="icon-info svg" title="" href="https://api.onlyoffice.com/editors/owncloud#masterKey" data-original-title="encryption:enable-master-key"></a>
 	</p>
 <?php } ?>
 	<div id="onlyofficeAddrSettings">
@@ -143,6 +142,15 @@ script("onlyoffice", "template");
 			<label for="onlyofficeCronChecker"><?php p($l->t("Enable background connection check to the editors")) ?></label>
 		</p>
 
+		<p>
+			<input type="checkbox" class="checkbox" id="onlyofficeEmailNotifications"
+				<?php if ($_["emailNotifications"]) { ?>checked="checked"<?php } ?> />
+			<label for="onlyofficeEmailNotifications"><?php p($l->t("Enable e-mail notifications")) ?></label>
+		</p>
+
+		<p class="onlyoffice-header"><?php p($l->t("Unknown author display name")) ?></p>
+		<p><input id="onlyofficeUnknownAuthor" value="<?php p($_["unknownAuthor"]) ?>" placeholder="" type="text"></p>
+
 		<p class="onlyoffice-header"><?php p($l->t("The default application for opening the format")) ?></p>
 		<div class="onlyoffice-exts">
 			<?php foreach ($_["formats"] as $format => $setting) { ?>
@@ -160,7 +168,6 @@ script("onlyoffice", "template");
 
 		<p class="onlyoffice-header">
 			<?php p($l->t("Open the file for editing (due to format restrictions, the data might be lost when saving to the formats from the list below)")) ?>
-				<a target="_blank" class="icon-info svg" title="" href="https://api.onlyoffice.com/editors/owncloud#editable" data-original-title="<?php p($l->t("View details")) ?>"></a>
 		</p>
 		<div class="onlyoffice-exts">
 			<?php foreach ($_["formats"] as $format => $setting) { ?>
@@ -179,7 +186,6 @@ script("onlyoffice", "template");
 
 		<h2>
 			<?php p($l->t("Editor customization settings")) ?>
-			<a target="_blank" class="icon-info svg" title="" href="https://api.onlyoffice.com/editors/config/editor/customization" data-original-title="<?php p($l->t("View details")) ?>"></a>
 		</h2>
 
 		<p>